Lots of fixes
This is great version to fall back to if needed. I was able to get firebase working and supporting tables. the game now works and correctly adds new users to firebase. Also, the game correctly goes to the intro scene and only asks the user to create a username once